refactor: AbstractTripComponent
