+++ src/components/trip/blank-trip.js
@@ -1,4 +1,4 @@
-import AbstractTripComponent from '../abstract-trip-component.js';
+import AbstractTripComponent from './abstract-trip-component.js';
 
 const createNoEventsTemplate = () => {
   return (
