+++ src/components/trip/events/collapsed-event.js
@@ -1,7 +1,7 @@
 import {eventTypesMap} from '../../../helpers/const.js';
 import {createOfferItemTemplate} from './offer-template.js';
 import {castTimeFormat, getFormatTime24H, getISOStringDate} from '../../../helpers/utils.js';
-import AbstractTripComponent from '../../abstract-trip-component.js';
+import AbstractTripComponent from '../abstract-trip-component.js';
 
 const SHOWING_OFFERS_COUNT = 3;
