+++ src/components/trip/events/editable-event.js
@@ -155,4 +155,9 @@ export default class EditableEvent extends AbstractComponent {
     this.getElement().querySelector(`form`)
       .addEventListener(`submit`, handler);
   }
+
+  setFavoritesButtonClickHandler(handler) {
+    this.getElement().querySelector(`.event__favorite-checkbox`)
+      .addEventListener(`click`, handler);
+  }
 }
